fix: 이미지 src 수정 및 date 유틸 shared 폴더로 이동
