+++ src/shared/utils/dateUtils.ts
@@ -28,3 +28,10 @@ export const isFutureDate = (checkDate: Date) => {
   targetDate.setHours(0, 0, 0, 0);
   return targetDate > today;
 };
+
+export const getTodayDateForForm = () => {
+  const now = new Date();
+  const offset = now.getTimezoneOffset();
+  const localToUTC = new Date(now.getTime() - offset * 60 * 1000);
+  return localToUTC.toISOString().split('T')[0];
+};
